refactor(playground): extract payload hydration helpers
